libm:gtests:ulp:Fixed ulp rounding off issue.

+++ gtests/acosh/acosh_callback.cc
@@ -44,12 +44,12 @@ double getFuncOp(double *data) {
   return LIBM_FUNC(acosh)(data[0]);
 }
 
-float getExpected(float *data) {
+double getExpected(float *data) {
   auto val = alm_mp_acoshf(data[0]);
   return val;
 }
 
-double getExpected(double *data) {
+long double getExpected(double *data) {
   auto val = alm_mp_acosh(data[0]);
   return val;
 }
