+++ gtests/tan/tan_callback.cc
@@ -44,12 +44,12 @@ double getFuncOp(double *data) {
   return LIBM_FUNC(tan)(data[0]);
 }
 
-float getExpected(float *data) {
+double getExpected(float *data) {
   auto val = alm_mp_tanf(data[0]);
   return val;
 }
 
-double getExpected(double *data) {
+long double getExpected(double *data) {
   auto val = alm_mp_tan(data[0]);
   return val;
 }
